Modularize files into HSMPRequest & HSMPResponse Move User class to its own header Move CreateRequest() to main.cpp
