+++ CMakeLists.txt
@@ -8,8 +8,10 @@ set(CMAKE_CXX_EXTENSIONS OFF)
 
 include_directories(${CMAKE_CURRENT_SOURCE_DIR}/src)
 
-add_executable(hsmp src/main.cpp)
+set(HSMP_SOURCES src/HSMP/HSMPRequest.cpp src/HSMP/HSMPResponse.cpp)
 
-add_executable(server src/server.cpp)
+add_executable(hsmp src/main.cpp ${HSMP_SOURCES})
 
-add_executable(client src/client.cpp)
+add_executable(server src/server.cpp ${HSMP_SOURCES})
+
+add_executable(client src/client.cpp ${HSMP_SOURCES})
