+++ src/client.cpp
@@ -7,7 +7,8 @@
 #include <sys/types.h>
 #include <unistd.h>
 #include <iostream>
-#include "HSMP.hpp"
+#include "HSMP/HSMPRequest.hpp"
+#include "HSMP/HSMPResponse.hpp"
 
 int main(void) {
   struct sockaddr_in stSockAddr;
